+++ tests/TestCase/FileArrayTest.php
@@ -33,9 +33,7 @@ class FileArrayTest extends TestCase
     protected $example = ['first', 'second', 'third', 'fourth', 'fifth'];
 
     /**
-     * Setup the test case, backup the static object values so they can be
-     * restored. Specifically backs up the contents of Configure and paths in
-     *  App if they have not already been backed up
+     * Called before every test method
      * @return void
      */
     public function setUp()